Add typechecking for RecursiveRecordTypeMerge.
The implementation brings with it intersection_with_key over HashMaps to
help with the type checking of records of records.

The implementation first checks that the Const values line up with the
LHS and RHS. Then checks that combining the records does not result in a
FieldCollision. It will finally return the shared Const type of the
arguments.