Make reqwest an optional dependency
This adds a feature to the Cargo.toml to be able to control if dhall depends on reqwest.
This makes it possible to build dhall-rust for any arch (not just wasm32) without depending on reqwest.

To disable reqwest as a dependency, we can now use `default-features = false` when adding the dhall dependency.

Closes #169